Crate imxrt_async_hal

Embedded, async Rust for i.MX RT processors

imxrt-async-hal brings async Rust support to NXP's i.MX RT processors. The crate includes awaitable peripherals and timers. Once the I/O completes or the timer elapses, an interrupt fires to wake the executor. By combining interrupt-driven peripherals with a single-threaded executor, you can write multiple, concurrent tasks for your embedded system.

The crate registers interrupt handlers to support async execution. The implementation registers interrupt handlers statically, using the cortex-m-rt interfaces. This means that your final program should also depend on cortex-m-rt, or bet at least cortex-m-rt compatible.

The crate does not include an executor, or any API for driving futures. You will need to select your own executor that supports a Cortex-M system. The executor should be thread safe, prepared to handle wakes from interrupt handlers.

See the project's examples to try this code on your hardware. This crate has been primarily developed using a Teensy 4 (i.MX RT 1062). It compiles for other i.MX RT chip variants.

Dependencies

A Rust installation; recommended installation using rustup. We support the latest, stable Rust toolchain.

The thumbv7em-none-eabihf Rust target, which may be installed using rustup: rustup target add thumbv7em-none-eabihf

The target is only necessary when building for an embedded system. The main crate should build and test on your host.

An embedded system with a compatible i.MX RT processor.

Feature flags

You're required to specify a feature that describes your i.MX RT chip variant. You may select only one chip feature.

The crate compiles for the following chips:

"imxrt1010" for i.MX RT 1010 variants

"imxrt1060" for i.MX RT 1060 variants

Each peripheral has it's own feature, which is enabled by default. However, you may want to disable some peripherals because you have your own interrupt-driven peripheral, and the interrupt handler that this crate provides causes a duplicate definition

To select peripherals, disable the crate's default features. Then, select one or more of the peripheral features from the table. The checkmarks indicate a chip's support for that peripheral.

Chip	"gpio"	"gpt"	"i2c"	"pipe"	"pit"	"spi"	"uart"

imxrt1010	✓	✓	✓	✓	✓	✓	✓

imxrt1060	✓	✓	✓	✓	✓	✓	✓

When developing a binary for your embedded system, you should enable this crate's "rt" feature. Otherwise, when developing libraries against the crate, you may skip the "rt" feature.

Examples

Each module contains a small getting started example to demonstrate driver setup and the driver API. For larger examples that run on hardware, see the imxrt-async-hal repository.

License

Licensed under either of

Apache License, Version 2.0

MIT License

at your option.

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use imxrt_ral as ral;

Modules

dma	pipe or spi or uart
Direct Memory Access (DMA) for async I/O

gpio	gpio
GPIOs

instance
Strongly-typed peripherals

iomuxc
Pad multiplexing and configuration

Structs

GPT	gpt
General purpose timers (GPT)

I2C	i2c
I2C Driver

PIT	pit
Periodic interrupt timer (PIT)

SPI	spi
Serial Peripheral Interface (SPI)

SPIPins	spi
Pins for a SPI device

UART	uart
UART Serial driver

Enums

I2CClockSpeed	i2c
I2C clock speed

I2CError	i2c
Errors propagated from an I2C device

SPIError	spi
Errors propagated from a SPI device

UARTError	uart
Errors propagated from a UART device