Module imxrt_async_hal::ccm

Clock control module (CCM)

The clocks and types exposed in ccm support clock control and peripheral clock gating. Use CCM::from_ral to acquire the clock roots and the CCM handle. Then, enable your clocks.

use imxrt_async_hal as hal;
use hal::{ccm, ral};

let ccm::CCM {
    mut handle,
    uart_clock,
    ..
} = ral::ccm::CCM::take().map(ccm::CCM::from_ral).unwrap();

let mut uart_clock = uart_clock.enable(&mut handle);

Clocks can enable peripheral clock gates, and they may be used in APIs that require you to first initialize clocks.

type UART2 = hal::instance::UART<hal::iomuxc::consts::U2>;
let mut lpuart2: UART2 = ral::lpuart::LPUART2::take().and_then(hal::instance::uart).unwrap();

// Enable the clock gate:
uart_clock.set_clock_gate(&mut lpuart2, ccm::ClockGate::On);

// Create the peripheral... see UART documentation for more information.

Structs

Handle
Handle to the CCM register block

Enums

ClockGate
Describes a clock gate setting

Type Definitions

CCM
Helper for a clock control module designed to the RAL interface.

I2CClock
An I2C clock that contorls RAL LPI2C timing

PerClock
A periodic clock that controls RAL PIT and GPT timings

SPIClock
A SPI clock that controls RAL LPSPI timing

UARTClock
A UART clock that controls RAL LPUART timing