Module imxrt_async_hal::dma

This is supported on crate features spi or uart only.

Direct Memory Access (DMA) for async I/O

DMA Channels power some asynchronous I/O operations. Use channels to acquire all of the processor's DMA channels. Then, use the Channels in APIs that require them. The implementation handles DMA receive and transfer operations, and ensures that the lifetime of your buffers is correct.

Structs

Channel
A DMA channel

Error
A wrapper around a DMA error status value

FullDuplex
A full-duplex DMA transfer from a single buffer

Rx
A DMA transfer that receives data from hardware

Tx
A DMA transfer that sends data to hardware

Enums

BandwidthControl
Throttles the amount of bus bandwidth consumed by the eDMA

Constants

CHANNEL_COUNT

Traits

Element
Describes a transferrable DMA element; basically, an unsigned integer of any size.

Functions

channels
Initialize and acquire the DMA channels

full_duplex
Perform a full-suplex DMA transfer using two DMA channels that read and write from a single buffer.

receive
Use a DMA channel to receive a buffer of elements from the source peripheral

transfer
Use a DMA channel to send a buffer of data to the destination peripheral