Module imxrt_async_hal::dma::pipe

This is supported on (crate features pipe or spi or uart) and crate feature pipe only.

Cross-task communication with DMA channels

(Note: this API is similar to typical 'channels' that you'll find in the Rust ecosystem. We use 'pipe' to disambiguate between this software channel and a hardware DMA channel.)

pipe provides a mechanism for sending data across tasks. The Sender half can send Copy data, and the Receiver half can receive that same data. The tasks use a DMA channel to transfer the data across tasks. Use new to create both halves of a pipe.

A Sender blocks until the Receiver is ready to receive data. Likewise, the Receiver blocks until the Sender is ready to send data. This creates a synchronization point for the two tasks. When the transfer completes, the data will have been transferred from the sender to the receiver.

The implementation does not guarantee any order for waking the two waiting tasks. That is, after the transfer completes, the sender task may wake before the receiver task; or, the receiver task may wake before the sender task.

To cancel a transfer, drop either the Sender or the Receiver. When one half is dropped, the remaining half will immediately return an Error::Cancelled. The remaining half can never be used again, as it will always, immediately return Error::Cancelled.

Example

Transmit an incrementing counter every 100ms using DMA channel 13. The sender is delayed by a GPT timer, which delays the receiver.

use imxrt_async_hal as hal;
use hal::{ccm::{CCM, ClockGate}, dma};
use hal::ral::{ccm, dma0, dmamux, gpt::GPT1};

let mut ccm = ccm::CCM::take().map(CCM::from_ral).unwrap();
let mut perclock = ccm.perclock.enable(&mut ccm.handle);
let (_, mut gpt, _) = GPT1::take()
    .map(|mut inst| {
        perclock.set_clock_gate_gpt(&mut inst, ClockGate::On);
        hal::GPT::new(inst, &mut perclock)
    })
    .unwrap();

let mut dma = dma0::DMA0::take().unwrap();
ccm.handle.set_clock_gate_dma(&mut dma, ClockGate::On);

let mut channels = dma::channels(
    dma,
    dmamux::DMAMUX::take().unwrap(),
);

let (mut tx, mut rx) = dma::pipe::new(channels[13].take().unwrap());
let sender = async {
    let mut counter: i32 = 0;
    loop {
        tx.send(&counter).await.unwrap();
        gpt.delay_us(100_000u32).await;
        counter = counter.wrapping_add(1);
    }
};

let receiver = async {
    loop {
        // Unblocks every 100ms, since that's the send rate.
        let counter = rx.receive().await.unwrap();
    }
};

block_on(futures::future::join(sender, receiver));

Structs

Receiver
The receiving half of a pipe

Sender
The sending half of a pipe

Functions

new
Create a pipe for sending and receiving data

Type Definitions

Result
Alias for a Result that might return an Error.