Struct imxrt_async_hal::dma::Channel

pub struct Channel { /* fields omitted */ }

This is supported on crate features pipe or spi or uart only.

A DMA channel

You should rely on your HAL to allocate Channels. If your HAL does not allocate channels, or if you're desigining the HAL, use new to create a new DMA channel.

You must always specify the source and destination transfer descriptors before enabling the transfer.

Implementations

impl Channel

pub fn set_bandwidth_control(&mut self, bandwidth: Option<BandwidthControl>)

Set the channel's bandwidth control

None disables bandwidth control (default setting)

Some(bwc) sets the bandwidth control to bwc

pub fn channel(&self) -> usize

Returns the DMA channel number

Channels are unique and numbered within the half-open range [0, CHANNEL_COUNT).

pub unsafe fn new(index: usize) -> Channel

Creates the DMA channel described by index

Safety

This will create a handle that may alias global, mutable state.

You must make sure that index describes a valid DMA channel for your system. If you're using this driver on a i.MX RT 1010 processor, you must make sure that index is less than 16.

Panics

Panics if index is greater than 32.

pub fn reset(&mut self)

Reset the transfer control descriptor owned by the DMA channel

reset should be called during channel initialization to put the channel into a known, good state.

pub unsafe fn set_source_transfer<E>(&mut self, transfer: &Transfer<E>) where
E: Element,

Prepare the source of a transfer; see Transfer for details.

Safety

User must ensure that the memory described by Transfer is valid for the lifetime of the DMA transaction.

pub unsafe fn set_destination_transfer<E>(&mut self, transfer: &Transfer<E>) where
E: Element,

Prepare the destination for a transfer; see Transfer for details.

Safety

User must ensure that the memory described by Transfer is valid for the lifetime of the DMA transaction.

pub fn set_minor_loop_bytes(&mut self, nbytes: u32)

Set the number of bytes to transfer per minor loop

Describes how many bytes we should transfer for each DMA service request.

pub fn set_minor_loop_elements<E>(&mut self, len: usize) where
E: Element,

Se the number of elements to move in each minor loop

Describes how many elements we should transfer for each DMA service request.

pub fn set_transfer_iterations(&mut self, iterations: u16)

Tells the DMA channel how many transfer iterations to perform

A 'transfer iteration' is a read from a source, and a write to a destination, with read and write sizes described by a minor loop. Each iteration requires a DMA service request, either from hardware or from software.

pub fn set_trigger_from_hardware(&mut self, source: Option<u32>)

Enable or disabling triggering from hardware

If source is Some(value), we trigger from hardware identified by the source identifier. If source is None, we disable hardware triggering.

pub fn set_always_on(&mut self)

Set this DMA channel as always on

Use set_always_on() so that the DMA multiplexer drives the transfer with no throttling. Specifically, an "always-on" transfer will not need explicit re-activiation between major loops.

Use an always-on channel for memory-to-memory transfers, so that you don't need explicit software re-activation to maintain the transfer. On the other hand, most peripheral transfers should not use an always-on channel, since the peripheral should control the data flow through explicit activation.

pub fn is_hardware_signaling(&self) -> bool

Returns true if the DMA channel is receiving a service signal from hardware

pub unsafe fn enable(&mut self)

Enable the DMA multiplexer request, which signals that the transfer is ready

Safety

This could initiate a DMA transaction that uses an invalid source or destination. Caller must ensure that the source and destination transfer descriptors are valid. See set_source_transfer and set_destination_transfer for more information.

pub fn disable(&mut self)

Disable the DMA channel, preventing any DMA transfers

pub fn is_interrupt(&self) -> bool

Returns true if this DMA channel generated an interrupt

pub fn clear_interrupt(&mut self)

Clear the interrupt flag from this DMA channel

pub fn set_disable_on_completion(&mut self, dreq: bool)

Enable or disable 'disable on completion'

'Disable on completion' lets the DMA channel automatically clear the request signal when it completes a transfer.

pub fn set_interrupt_on_completion(&mut self, intr: bool)

Enable or disable interrupt generation when the transfer completes

You're responsible for registering your interrupt handler.

pub fn is_complete(&self) -> bool

Indicates if the DMA transfer has completed

pub fn clear_complete(&mut self)

Clears completion indication

pub fn is_error(&self) -> bool

Indicates if the DMA channel is in an error state

pub fn clear_error(&mut self)

Clears the error flag

pub fn is_active(&self) -> bool

Indicates if this DMA channel is actively transferring data

pub fn is_enabled(&self) -> bool

Indicates if this DMA channel is enabled

pub fn error_status(&self) -> ErrorStatus

Returns the value from the global error status register

It may reflect the last channel that produced an error, and that may not be related to this channel.

pub unsafe fn start(&mut self)

Start a DMA transfer

start() should be used to request service from the DMA controller. It's necessary for in-memory DMA transfers. Do not use it for hardware-initiated DMA transfers. DMA transfers that involve hardware will rely on the hardware to request DMA service.

Flag is automatically cleared by hardware after it's asserted.

Safety

This could initiate a DMA transaction that uses an invalid source or destination. Caller must ensure that the source and destination transfer descriptors are valid. See set_source_transfer and set_destination_transfer for more information.