Struct imxrt_async_hal::gpio::GPIO

pub struct GPIO<P, D> { /* fields omitted */ }

This is supported on crate feature gpio only.

A wrapper around an i.MX RT processor pad, supporting simple I/O

Use new with a pad from the iomuxc module. All GPIOs start in the input state. Use output to become an output pin.

use imxrt_async_hal as hal;
use hal::gpio::GPIO;

let pads = hal::iomuxc::new(hal::ral::iomuxc::IOMUXC::take().unwrap());
let mut led = GPIO::new(pads.b0.p03);
assert!(!led.is_set());
let mut led = led.output();
led.set();

When using a GPIO as an input, you can wait for transitions using wait_for.

Implementations

impl<P> GPIO<P, Input> where
P: Pin,

pub fn new(pin: P) -> Self

Create a GPIO from a pad that supports a GPIO configuration

All pads may be used as a GPIO, so new should work with every iomuxc pad.

use imxrt_async_hal as hal;
use hal::gpio::GPIO;

let pads = hal::iomuxc::new(hal::ral::iomuxc::IOMUXC::take().unwrap());
let input_pin = GPIO::new(pads.b0.p03);

pub fn output(self) -> GPIO<P, Output>

Transition the GPIO from an input to an output

pub fn is_set(&self) -> bool

Returns true if this input pin is high

pub fn wait_for(&mut self, trigger: Trigger) -> Interrupt<'_, P>

Sets the trigger for the input GPIO, and await for the input event.

use imxrt_async_hal as hal;
use hal::gpio::{GPIO, Trigger};

let pads = hal::iomuxc::new(hal::ral::iomuxc::IOMUXC::take().unwrap());
let mut input_pin = GPIO::new(pads.ad_b1.p02);
// ...
input_pin.wait_for(Trigger::RisingEdge).await;

impl<P> GPIO<P, Output> where
P: Pin,

pub fn input(self) -> GPIO<P, Input>

Transition the pin from an output to an input

pub fn set(&mut self)

Drive the GPIO high

pub fn clear(&mut self)

Drive the GPIO low

pub fn is_set(&self) -> bool

Returns true if the pin is driving high

pub fn toggle(&mut self)

Alternate the state of the pin

Using toggle will be more efficient than checking is_set and then selecting the opposite state.