Module imxrt_async_hal::i2c

I2C driver, types, and futures

The I2C driver utilizes the internal transmit and receive FIFOs to send and receive data. When the transmit buffer is nearly full, the driver yields. When the transmit buffer can support more data, an I2C interrupt fires and wakes the executor. This cycle continues until all of the caller's data is transmitted.

When the receive buffer does not have any data, but the caller is awaiting data, the drier yields. Once there's at least one byte in the receive buffer, an I2C interrupt fires and wakers the executor. This cycle continues until all of the caller's receive buffer is filled.

The driver also yields when waiting for stop and repeated start conditions.

The I2C clock speed is unspecified out of construction. Use set_clock_speed to select a valid I2C clock speed.

The RAL instances are available in ral::lpi2c.

Pin configuration

You may need to configure the SCL and SDA pins to support your clock speed and data rate. The snippet below provides one possible configuration that supports both 100KHz and 400KHz I2C clock speeds.

use imxrt_async_hal as hal;
use hal::iomuxc;

const PINCONFIG: iomuxc::Config = iomuxc::Config::zero()
    .set_open_drain(iomuxc::OpenDrain::Enabled)
    .set_slew_rate(iomuxc::SlewRate::Fast)
    .set_drive_strength(iomuxc::DriveStrength::R0_4)
    .set_speed(iomuxc::Speed::Fast)
    .set_pull_keep(iomuxc::PullKeep::Enabled)
    .set_pull_keep_select(iomuxc::PullKeepSelect::Pull)
    .set_pullupdown(iomuxc::PullUpDown::Pullup22k);

Example

Prepare the I2C3 peripheral at 400KHz. Note that this example does not demonstrate how to set up the I2C peripheral clocks, and enable clock gates.

use imxrt_async_hal as hal;
use hal::{
    iomuxc, I2C, I2CClockSpeed,
    ral::{self, ccm::CCM, iomuxc::IOMUXC, lpi2c::LPI2C3},
};
const SOURCE_CLOCK_HZ: u32 = 24_000_000;
const SOURCE_CLOCK_DIVIDER: u32 = 3;

let ccm = CCM::take().unwrap();
// LPI2C clock selection: 24MHz XTAL, divide by 3
ral::modify_reg!(ral::ccm, ccm, CSCDR2, LPI2C_CLK_SEL: 1, LPI2C_CLK_PODF: SOURCE_CLOCK_DIVIDER - 1);
// LPI2C3 clock gate on
ral::modify_reg!(ral::ccm, ccm, CCGR2, CG5: 0b11);

let mut pads = IOMUXC::take()
    .map(iomuxc::new)
    .unwrap();

iomuxc::configure(&mut pads.ad_b1.p07, PINCONFIG);
iomuxc::configure(&mut pads.ad_b1.p06, PINCONFIG);

let mut i2c3 = LPI2C3::take().and_then(hal::instance::i2c).unwrap();

let mut i2c = I2C::new(i2c3, pads.ad_b1.p07, pads.ad_b1.p06);
i2c.set_clock_speed(I2CClockSpeed::KHz400, SOURCE_CLOCK_HZ / SOURCE_CLOCK_DIVIDER).unwrap();

let output = [1, 2, 3, 4];
let mut input = [0; 7];
i2c.write_read(DEVICE_ADDRESS, &output, &mut input).await.unwrap();

Structs

I2C	i2c
The I2C driver instance

Read
An I2C read future

Write
An I2C write future

WriteRead
An I2C write-read future

Enums

ClockSpeed	i2c
I2C clock speed

Error	i2c
Errors propagated from an I2C device

State
I2C polling state