Module imxrt_async_hal::instance

Strongly-typed peripherals

Peripheral instances from the RAL do not carry their peripheral identification in the type system. For instance, an LPUART2 peripheral and an LPUART3 peripheral are represented by the same Rust type, ral::lpuart::Instance.

However, the iomuxc APIs work with strongly-typed peripheral instances, which are identified by a type-level constant. This interface expects LPUART2 and LPUART3 to be unique types. To bridge these APIs, and ensure that your peripheral instances work with your pin selections, use the instance interface.

An Instance wraps a RAL peripheral instance, requiring that the peripheral instance matches its type-level constant. The interface ensures that your RAL instance matches the type-level constant:

use imxrt_async_hal as hal;
use hal::{instance, iomuxc::consts};
use hal::ral::lpuart;

// This number... ----------------v
let inst: instance::UART<consts::U2> =
    lpuart::LPUART2::take().and_then(instance::uart).unwrap();
//                ^----- ...matches here, so the unwrap is safe!

A mismatch between the expected instance and the RAL instance will return None, or panic whe unwrapped:

// This number... ----------------v
let inst: instance::UART<consts::U2> =
    lpuart::LPUART3::take().and_then(instance::uart).unwrap();
//                ^----- ...doesn't match! panic!

Typically, you may elide the types, since the peripheral APIs will match the expected types. See the documentation of your peripheral for examples.

Structs

Instance
A strongly-typed RAL instance

Traits

Inst
A trait implemented on RAL instances

Functions

i2c	i2c
Specify an I2C instance

spi	spi
Specify a SPI instance

uart	uart
Specify a UART instance

Type Definitions

I2C	i2c
Alias for an Instance around a ral::lpi2c::Instance

SPI	spi
Alias for an Instance around a ral::lpspi::Instance

UART	uart
Alias for an Instance around a ral::lpuart::Instance