Module imxrt_async_hal::iomuxc

Pad multiplexing and configuration

The iomuxc module is a re-export of the imxrt-iomuxc crate. It combines the i.MX RT processor-specific components with the imxrt-iomuxc general API. It then adds a safe function, take, which lets you convert the RAL's iomuxc::Instance into all of the processor Pads.

use imxrt_async_hal as hal;
use hal::{ral::iomuxc::IOMUXC, iomuxc};

let pads = iomuxc::new(IOMUXC::take().unwrap());

Pads can then be used in peripheral-specific APIs.

Modules

consts
Type-level constants and traits

gpio
GPIO pad configuration

i2c
I2C pad configuration

pads

pwm
PWM pad configuration

spi
SPI pad configurations

uart
UART pad configuration

Structs

Config
A configuration capable of compile-time, const configuration:

Daisy
A daisy selection

ErasedPad
A pad that has its type erased

Pad
An i.MXT RT pad

WrongPadError
An error that indicates the conversion from an ErasedPad to a strongly-typed pad failed.

Enums

DriveStrength
Drive strength

Hysteresis
The hysteresis (HYS) bit controls whether a pin acts as a Schmitt trigger, which is a comparator remembering its last input state (hysteresis).

OpenDrain
Open Drain Enable Field

PullKeep
Enable or disable the pull / keeper functionality

PullKeepSelect
Control signal to enable internal pull-up/down resistors or pad keeper functionality.

PullUpDown
Controls signals to select pull-up or pull-down internal resistance strength.

SlewRate
Slew Rate

Speed
Sets electrical characteristics of a pin in a given frequency range

Functions

configure
Applies the configuration config for the supplied pad

new	imxrt1010 or imxrt1060
Turn the IOMUXC instance into pads