Struct imxrt_async_hal::iomuxc::pads::ad_b1::Pads

pub struct Pads {
    pub p00: Pad<AD_B1, UTerm>,
    pub p01: Pad<AD_B1, UInt<UTerm, B1>>,
    pub p02: Pad<AD_B1, UInt<UInt<UTerm, B1>, B0>>,
    pub p03: Pad<AD_B1, UInt<UInt<UTerm, B1>, B1>>,
    pub p04: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B0>, B0>>,
    pub p05: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B0>, B1>>,
    pub p06: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B1>, B0>>,
    pub p07: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B1>, B1>>,
    pub p08: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>>,
    pub p09: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B1>>,
    pub p10: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B1>, B0>>,
    pub p11: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B1>, B1>>,
    pub p12: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B0>>,
    pub p13: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B1>>,
    pub p14: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B1>, B0>>,
    pub p15: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B1>, B1>>,
}

Pads with the prefix 'AD_B1'

Fields
p00: Pad<AD_B1, UTerm>p01: Pad<AD_B1, UInt<UTerm, B1>>p02: Pad<AD_B1, UInt<UInt<UTerm, B1>, B0>>p03: Pad<AD_B1, UInt<UInt<UTerm, B1>, B1>>p04: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B0>, B0>>p05: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B0>, B1>>p06: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B1>, B0>>p07: Pad<AD_B1, UInt<UInt<UInt<UTerm, B1>, B1>, B1>>p08: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>>p09: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B1>>p10: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B1>, B0>>p11: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B1>, B1>>p12: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B0>>p13: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B1>>p14: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B1>, B0>>p15: Pad<AD_B1, UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B1>, B1>>
Implementations

impl Pads

pub const unsafe fn new() -> Pads

Take all pads from this group

Safety

You may safely call this once to acquire all of the pads. Subsequent calls may return pads that are mutably aliased elsewhere. Consider calling new() at the start of your program.

Know that the top-level Pads::new() will call this new().

pub fn erase(self) -> [ErasedPad; 16]

Erase all of the pads

The return type is an array, where the index indicates the pad offset from the start of the group. For example, AD_B0_03 would be referenced as erased_pads[3].

See ErasedPad for more information.