Struct imxrt_async_hal::iomuxc::Pad

pub struct Pad<Base, Offset> { /* fields omitted */ }

An i.MXT RT pad

The Base is the pad tag, like AD_B0. The Offset is the constant (type) that describes the pad number.

Pads have no size.

Implementations

impl<Base, Offset> Pad<Base, Offset>

pub const unsafe fn new() -> Pad<Base, Offset>

Creates a handle to the pad

Safety

new() may be called anywhere, by anyone. This could lead to multiple objects that mutate the same memory. Consider calling new() once, near startup, then passing objects and references throughout your program.

impl<Base, Offset> Pad<Base, Offset> where
Base: Base,
Offset: Unsigned,

pub fn erase(self) -> ErasedPad

Erase the pad's type, returning an ErasedPad