Struct imxrt_async_hal::I2C

pub struct I2C<SCL, SDA> { /* fields omitted */ }

This is supported on crate feature i2c only.

I2C Driver

The I2C driver utilizes the internal transmit and receive FIFOs to send and receive data. When the transmit buffer is nearly full, the driver yields. When the transmit buffer can support more data, an I2C interrupt fires and wakes the executor. This cycle continues until all of the caller's data is transmitted.

When the receive buffer does not have any data, but the caller is awaiting data, the drier yields. Once there's at least one byte in the receive buffer, an I2C interrupt fires and wakers the executor. This cycle continues until all of the caller's receive buffer is filled.

The driver also yields when waiting for stop and repeated start conditions.

The I2C clock speed is unspecified out of construction. Use set_clock_speed to select a valid I2C clock speed.

The RAL instances are available in ral::lpi2c.

Pin configuration

You may need to configure the SCL and SDA pins to support your clock speed and data rate. The snippet below provides one possible configuration that supports both 100KHz and 400KHz I2C clock speeds.

use imxrt_async_hal as hal;
use hal::iomuxc;

const PINCONFIG: iomuxc::Config = iomuxc::Config::zero()
    .set_open_drain(iomuxc::OpenDrain::Enabled)
    .set_slew_rate(iomuxc::SlewRate::Fast)
    .set_drive_strength(iomuxc::DriveStrength::R0_4)
    .set_speed(iomuxc::Speed::Fast)
    .set_pull_keep(iomuxc::PullKeep::Enabled)
    .set_pull_keep_select(iomuxc::PullKeepSelect::Pull)
    .set_pullupdown(iomuxc::PullUpDown::Pullup22k);

Example

Prepare the I2C3 peripheral at 400KHz, using Teensy pins 16 and 17.

use imxrt_async_hal as hal;
use hal::{
    ccm, iomuxc, I2C, I2CClockSpeed,
    ral::{ccm::CCM, iomuxc::IOMUXC, lpi2c::LPI2C3},
};

let mut pads = IOMUXC::take()
    .map(iomuxc::new)
    .unwrap();

iomuxc::configure(&mut pads.ad_b1.p07, PINCONFIG);
iomuxc::configure(&mut pads.ad_b1.p06, PINCONFIG);

let mut ccm = CCM::take().map(ccm::CCM::from_ral).unwrap();
let mut i2c_clock = ccm.i2c_clock.enable(&mut ccm.handle);
let mut i2c3 = LPI2C3::take().and_then(hal::instance::i2c).unwrap();
i2c_clock.set_clock_gate(&mut i2c3, ccm::ClockGate::On);

let mut i2c = I2C::new(i2c3, pads.ad_b1.p07, pads.ad_b1.p06, &i2c_clock);
i2c.set_clock_speed(I2CClockSpeed::KHz400).unwrap();

let output = [1, 2, 3, 4];
let mut input = [0; 7];
i2c.write_read(DEVICE_ADDRESS, &output, &mut input).await.unwrap();

Implementations

impl<SCL, SDA, M> I2C<SCL, SDA> where
M: Unsigned,
SCL: Pin<Signal = SCL, Module = M>,
SDA: Pin<Signal = SDA, Module = M>,

pub fn new(i2c: I2C<M>, scl: SCL, sda: SDA, clock: &I2CClock) -> Self

Create an I2C driver from an I2C instance and a pair of I2C pins

The I2C clock speed of the returned I2C driver is unspecified and may not be valid. Use set_clock_speed to select a valid I2C clock speed.

impl<SCL, SDA> I2C<SCL, SDA>

pub fn release(self) -> (Instance, SCL, SDA)

Release the I2C peripheral components

pub fn set_clock_speed(&mut self, clock_speed: ClockSpeed) -> Result<(), Error>

Set the I2C clock speed

If there is an error, error variant is crate::i2c::Error::ClockSpeed.

pub async fn write_read<'_, '_, '_>(
&'_ mut self,
address: u8,
output: &'_ [u8],
input: &'_ mut [u8]
) -> Result<(), Error>

Perform a write-read to an I2C device identified by address

Sends output, generates a repeated start, then awaits the I2C device to send enough data for input.

pub async fn write<'_, '_>(
&'_ mut self,
address: u8,
buffer: &'_ [u8]
) -> Result<(), Error>

Perform an I2C write, sending buffer to the I2C device identified by address

pub async fn read<'_, '_>(
&'_ mut self,
address: u8,
buffer: &'_ mut [u8]
) -> Result<(), Error>

Request a buffer of data from an I2C device identified by address