Struct imxrt_async_hal::SPI

pub struct SPI<Pins> { /* fields omitted */ }

This is supported on crate feature spi only.

Serial Peripheral Interface (SPI)

A SPI peripheral uses DMA for asynchronous I/O. Using up to two DMA channels, SPI peripherals can perform SPI device reads, writes, and full-duplex transfers with u8 and u16 elements.

The SPI serial clock speed after construction is unspecified. Use set_clock_speed to choose your SPI serial clock speed.

The RAL instances are available in ral::lpspi.

Example

Perform a full-duplex SPI transfer of four u16s using SPI4.

use imxrt_async_hal as hal;
use hal::{dma, instance, iomuxc, SPI, SPIPins};
use hal::ral::{self,
    ccm::CCM, dma0::DMA0, dmamux::DMAMUX,
    iomuxc::IOMUXC, lpspi::LPSPI4,
};

// Effective LPSPI source clock (PLL2)
const SOURCE_CLOCK_HZ: u32 = 528_000_000;
// Any divider for the source clock
const SOURCE_CLOCK_DIVIDER: u32 = 5;

let pads = IOMUXC::take().map(iomuxc::new).unwrap();

let ccm = CCM::take().unwrap();
// Prepare SPI clocks
ral::modify_reg!(
    ral::ccm,
    ccm,
    CBCMR,
    LPSPI_CLK_SEL: LPSPI_CLK_SEL_2, /* PLL2 */
    LPSPI_PODF: SOURCE_CLOCK_DIVIDER - 1
);
// LPSPI4 clock gate on
ral::modify_reg!(ral::ccm, ccm, CCGR1, CG3: 0b11);
// DMA clock on
ral::modify_reg!(ral::ccm, ccm, CCGR5, CG3: 0b11);

let dma = DMA0::take().unwrap();
let mut channels = dma::channels(
    dma,
    DMAMUX::take().unwrap(),
);

let spi_pins = SPIPins {
    sdo: pads.b0.p02,
    sdi: pads.b0.p01,
    sck: pads.b0.p03,
    pcs0: pads.b0.p00,
};
let spi4 = LPSPI4::take().and_then(instance::spi).unwrap();
let mut spi = SPI::new(
    spi_pins,
    spi4,
    (channels[8].take().unwrap(), channels[9].take().unwrap()),
);

spi.set_clock_speed(1_000_000, SOURCE_CLOCK_HZ / SOURCE_CLOCK_DIVIDER).unwrap();

let mut buffer = [1, 2, 3, 4];
// Transmit the u16 words in buffer, and receive the reply into buffer.
spi.full_duplex_u16(&mut buffer).await;

Implementations

impl<SDO, SDI, SCK, PCS0, M> SPI<Pins<SDO, SDI, SCK, PCS0>> where
SDO: Pin<Module = M, Signal = SDO>,
SDI: Pin<Module = M, Signal = SDI>,
SCK: Pin<Module = M, Signal = SCK>,
PCS0: Pin<Module = M, Signal = PCS0>,
M: Unsigned,

pub fn new(
pins: Pins<SDO, SDI, SCK, PCS0>,
spi: SPI<M>,
channels: (Channel, Channel)
) -> Self

Create a SPI from a set of pins, a SPI peripheral instance, and two DMA channels

See the instance module for more information on SPI peripheral instances.

The clock speed is unspecified. Make sure you change your clock speed with set_clock_speed.

pub fn release(
self
) -> (Pins<SDO, SDI, SCK, PCS0>, Instance, (Channel, Channel))

Return the pins, instance, and DMA channels that are used in this SPI driver

impl<Pins> SPI<Pins>

pub fn set_clock_speed(
&mut self,
hz: u32,
source_clock_hz: u32
) -> Result<(), Error>

Set the SPI master clock speed

Consider calling set_clock_speed after creating a SPI, since the clock speed after construction is unspecified.

If an error occurs, it's an crate::spi::Error::ClockSpeed.

pub async fn read_u8<'_, '_>(
&'_ mut self,
buffer: &'_ mut [u8]
) -> Result<usize, Error>

Await for a u8 buffer of data from a SPI device

Blocks until buffer is filled. Returns the number of bytes placed in buffer.

pub async fn read_u16<'_, '_>(
&'_ mut self,
buffer: &'_ mut [u16]
) -> Result<usize, Error>

Await for a u16 buffer of data from a SPI device

Blocks until buffer is filled. Returns the number of bytes placed in buffer.

pub async fn write_u8<'_, '_>(
&'_ mut self,
buffer: &'_ [u8]
) -> Result<usize, Error>

Transmit a buffer of bytes to a SPI device

Blocks until the contents of buffer have been transferred from the host controller. Returns the number of bytes written.

pub async fn write_u16<'_, '_>(
&'_ mut self,
buffer: &'_ [u16]
) -> Result<usize, Error>

Transmit a buffer of u16s to a SPI device

Blocks until the contents of buffer have been transferred from the host controller. Returns the number of bytes written.

pub async fn full_duplex_u8<'_, '_>(
&'_ mut self,
buffer: &'_ mut [u8]
) -> Result<usize, Error>

Transfer bytes from buffer while simultaneously receiving bytes into buffer

Each transferred element from the buffer is replaced by an element read from the SPI device. Returns the number of elements sent and received.

pub async fn full_duplex_u16<'_, '_>(
&'_ mut self,
buffer: &'_ mut [u16]
) -> Result<usize, Error>

Transfer u16 words from buffer while simultaneously receiving u16 words into buffer

Each transferred element from the buffer is replaced by an element read from the SPI device. Returns the number of elements sent and received.